requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
llama-index
openai